refactor : reservation 웹 패키지 구조 개선
