+++ mealkitary-api/src/test/kotlin/com/mealkitary/reservation/web/PayReservationControllerTest.kt
@@ -1,8 +1,8 @@
-package com.mealkitary.reservation.adapter.input.web
+package com.mealkitary.reservation.web
 
 import com.mealkitary.WebIntegrationTestSupport
 import com.mealkitary.common.exception.EntityNotFoundException
-import com.mealkitary.reservation.adapter.input.web.request.PayReservationWebRequest
+import com.mealkitary.reservation.web.request.PayReservationWebRequest
 import io.mockk.every
 import org.springframework.http.MediaType
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post
